+++ .gitignore
@@ -1,1 +1,3 @@
 node_modules/
+lib/database/config.js
+npm-debug.log
