+++ lib/database/config.js.example
@@ -1,0 +1,6 @@
+module.exports = {
+  host: 'localhost',
+  user: 'root',
+  password: 'my-secret-pw',
+  database: 'exampleeeee'
+};
