+++ package.json
@@ -9,6 +9,7 @@
   "author": "Sinterklaas",
   "license": "ISC",
   "dependencies": {
+    "body-parser": "^1.15.2",
     "express": "^4.14.0",
     "mysql": "^2.12.0"
   }
